fix: fix clone environment lack of computed attributes
